+++ argilla-frontend/components/features/annotation/container/fields/text-field/TextField.vue
@@ -18,6 +18,7 @@
     </div>
     <div :id="`fields-content-${name}`" class="content-area --body1">
       <div :class="classes" v-if="!useMarkdown" v-html="fieldText" />
+      <Sandbox v-else-if="isHTML" :fieldText="fieldText" />
       <RenderMarkdownBaseComponent v-else :markdown="fieldText" />
       <template>
         <style :key="name" scoped>
@@ -60,6 +61,9 @@ export default {
     classes() {
       return this.$language.isRTL(this.fieldText) ? "--rtl" : "--ltr";
     },
+    isHTML() {
+      return /<([A-Za-z][A-Za-z0-9]*)\b[^>]*>(.*?)<\/\1>/.test(this.fieldText);
+    },
   },
   setup(props) {
     return useTextFieldViewModel(props);
@@ -74,8 +78,9 @@ export default {
   flex-direction: column;
   gap: $base-space;
   padding: 2 * $base-space;
-  background: palette(grey, 800);
+  background: var(--bg-field);
   border-radius: $border-radius-m;
+  border: 1px solid var(--bg-opacity-2);
   &:hover {
     #{$this}__copy-button {
       opacity: 1;
@@ -86,7 +91,7 @@ export default {
     align-items: center;
     justify-content: space-between;
     gap: $base-space;
-    color: $black-87;
+    color: var(--fg-primary);
   }
   .content-area {
     white-space: pre-wrap;