+++ argilla-frontend/translation/es.js
@@ -38,6 +38,19 @@ export default {
   taskDistributionTooltip:
     "Una tarea se completa cuando todos los \nregistros tienen el número mínimo \nde respuestas entregadas",
   noAnnotationGuidelines: "Este dataset no tiene guía de anotación",
+  noRecordsMessages: {
+    datasetEmptyForAnnotator:
+      "El dataset está vacío. Pide a un administrador que suba registros y vuelve pronto.",
+    datasetEmptyForAdmin:
+      "El dataset está vacío. Puedes agregar registros usando el SDK de Python, consulta la <a href='https://docs.argilla.io/latest/how_to_guides/record/'>documentación</a> sobre cómo agregar registros.",
+    taskDistributionCompleted: "🎉 ¡La tarea está completada!",
+    noSubmittedRecords: "Aún no has entregado ningún registro",
+    noRecordsFound:
+      "No tienes registros {status} que coincidan con tu búsqueda",
+    noRecords: "No tienes registros {status}",
+    noPendingRecordsToAnnotate: "🎉 No tienes registros pendientes para anotar",
+    noDraftRecordsToReview: "No tienes ningún borrador para revisar",
+  },
   breadcrumbs: {
     home: "Inicio",
     datasetSettings: "Configuración",
@@ -49,11 +62,11 @@ export default {
     pending: "Pendiente",
   },
   recordStatus: {
-    pending: "Pendiente",
-    draft: "Borrador",
-    discarded: "Descartado",
-    submitted: "Entregado",
-    validated: "Validado",
+    pending: "pendiente | pendientes",
+    draft: "borrador | borradores",
+    discarded: "descartado | descartados",
+    submitted: "entregado | entregados",
+    validated: "validado | validados",
     completedTooltip:
       "El registro está completo, tiene el número \nmínimo de respuestas",
   },
@@ -123,7 +136,7 @@ export default {
     select_to_annotate: "Seleccionar todo",
     pageSize: "Tamaño de página",
     selectAllResults: "Selecciona todos los registros coincidentes {total}",
-    haveSelectedRecords: "Has seleccionado todos los registros {Total}",
+    haveSelectedRecords: "Has seleccionado todos los registros {total}",
     actionConfirmation: "Confirmación de acción en bloque",
     actionConfirmationText:
       "Esta acción afectará a {total} registros, ¿Quiere continuar?",
@@ -164,7 +177,7 @@ export default {
   suggestion: {
     agent: "\nagente: {agent}",
     score: "\npuntuación: {score}",
-    tooltip: "Esta pregunta contiene una sugerencia {agente} {stork}",
+    tooltip: "Esta pregunta contiene una sugerencia {agent} {score}",
     filter: {
       value: "Valores de sugerencia",
       score: "Puntuación",
@@ -196,7 +209,7 @@ export default {
     passwordDescription: "Introduce tu contraseña",
     claim: "Trabaja en equipo con tus datos.</br>Perfecciona tus modelos.",
     hf: {
-      title: "Bienvenido a {Space}",
+      title: "Bienvenido a {space}",
       subtitle:
         "Únete a <strong>{user}</strong> para construir mejores datasets para IA",
     },