+++ argilla/src/argilla/records/_io/_datasets.py
@@ -19,10 +19,11 @@ from datasets import Dataset as HFDataset
 from datasets import IterableDataset, Image
 
 from argilla.records._io._generic import GenericIO
-from argilla._helpers._media import pil_to_data_uri
+from argilla._helpers._media import pil_to_data_uri, uncast_image
 
 if TYPE_CHECKING:
     from argilla.records import Record
+    from argilla.datasets import Dataset
 
 
 class HFDatasetsIO:
@@ -39,7 +40,7 @@ class HFDatasetsIO:
         return isinstance(dataset, HFDataset)
 
     @staticmethod
-    def to_datasets(records: List["Record"]) -> HFDataset:
+    def to_datasets(records: List["Record"], dataset: "Dataset") -> HFDataset:
         """
         Export the records to a Hugging Face dataset.
 
@@ -47,8 +48,11 @@ class HFDatasetsIO:
             The dataset containing the records.
         """
         record_dicts = GenericIO.to_dict(records, flatten=True)
-        dataset = HFDataset.from_dict(record_dicts)
-        return dataset
+        hf_dataset = HFDataset.from_dict(record_dicts)
+        image_fields = HFDatasetsIO._get_image_fields(schema=dataset.schema)
+        if image_fields:
+            hf_dataset = HFDatasetsIO._cast_uris_as_images(hf_dataset=hf_dataset, columns=image_fields)
+        return hf_dataset
 
     @staticmethod
     def _record_dicts_from_datasets(dataset: HFDataset) -> List[Dict[str, Union[str, float, int, list]]]:
@@ -71,6 +75,18 @@ class HFDatasetsIO:
         for example in dataset:
             record_dicts.append(example)
         return record_dicts
+
+    @staticmethod
+    def _get_image_fields(schema: Dict) -> List[str]:
+        """Get the names of the Argilla fields that contain image data.
+
+        Parameters:
+            dataset (Dataset): The dataset to check.
+
+        Returns:
+            List[str]: The names of the Argilla fields that contain image data.
+        """
+        return [field_name for field_name, field in schema.items() if field.type == "image"]
 
     @staticmethod
     def _get_image_features(dataset: "HFDataset") -> List[str]:
@@ -102,6 +118,35 @@ class HFDatasetsIO:
         def batch_fn(batch):
             data_uris = [pil_to_data_uri(sample) for sample in batch]
             return {unique_identifier: data_uris}
+
+        for column in columns:
+            hf_dataset = hf_dataset.map(
+                function=batch_fn,
+                with_indices=False,
+                batched=True,
+                input_columns=[column],
+                remove_columns=[column],
+            )
+            hf_dataset = hf_dataset.rename_column(original_column_name=unique_identifier, new_column_name=column)
+
+        return hf_dataset
+
+    @staticmethod
+    def _cast_uris_as_images(hf_dataset: "HFDataset", columns: List[str]) -> "HFDataset":
+        """Cast the image features in the Hugging Face dataset as PIL images.
+
+        Parameters:
+            hf_dataset (HFDataset): The Hugging Face dataset to cast.
+            columns (List[str]): The names of the columns containing the image features.
+
+        Returns:
+            HFDataset: The Hugging Face dataset with image features cast as PIL images.
+        """
+        unique_identifier = uuid4().hex
+
+        def batch_fn(batch):
+            images = [uncast_image(sample) for sample in batch]
+            return {unique_identifier: images}
 
         for column in columns:
             hf_dataset = hf_dataset.map(